pyproject.toml
[project]
name = "neuro-interview"
version = "0.1.0"
description = ""
authors = [
    {name = "hhhscvx",email = "hhhscvx@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "faster-whisper (>=1.2.0,<2.0.0)",
    "torch (==2.1.2)",
    "numpy (>=1.22,<1.24)",
    "torchaudio (==2.1.2)",
    "scikit-learn (==1.4.2)",
    "librosa (==0.10.2.post1)",
    "soundfile (==0.12.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)"
]

[tool.poetry.dependencies]
torch = {source = "pypi"}
torchaudio = {source = "pypi"}
